requirements.txt
asgiref
certifi
charset-normalizer 
colorama
crispy-bootstrap5  
defusedxml
distlib
Django
django-cors-headers
django-crispy-forms
django-environ     
django-filter      
djangorestframework
filelock
httpie
idna
Markdown
markdown-it-py
mdurl
MouseInfo
multidict
pillow
platformdirs
PyAutoGUI
PyGetWindow
Pygments
pyjokes
PyMsgBox
pyperclip
PyRect
PyScreeze
PySocks
pytweening
requests
requests-toolbelt
rich
setuptools
sqlparse
tzdata
urllib3
virtualenv